+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     - Swarm Pokemon show up in the Pokedex correctly
     - Implemented Birch's book for active swarms
 - Increased the variety of wild Pokemon in some areas
+- Added a list to the Pokemon's menu that shows any field moves it can use that can be used from the overworld.
+    - This fixes the issue that made Regirock unobtainable due to being locked behind Rock Smash.
 
 ## 0.1.0
 - Initial game release